refactor lut and histogram
